requirements.txt
pillow
requests
requests[socks]
pysocks
flask
flask-cors
pyyaml
aiohttp
aiohttp_socks
pynat